+++ game/js/scenes/endScene.js
@@ -1,0 +1,20 @@
+class EndScene extends Phaser.Scene {
+
+    constructor()
+    {
+        super({ key:'endScene', active: true});
+    }
+
+    preload()
+    {
+
+    }
+
+    create()
+    {
+
+    }
+
+};
+
+export default EndScene;
